requirements.txt
pandas
numpy
yfinance
ta
python-telegram-bot